added extra template and post
