+++ src/routes/index.ts
@@ -668,17 +668,8 @@ router.route(config.UPDATE_REVIEW_STATEMENT_BEFORE_SUBMITTING_URL)
     companyAuthentication,
     navigation.hasOverseasEntity
   )
-  .get(updateReviewStatement.get);
-// .post(...validator.registrableBeneficialOwner, checkValidations, noChangeRegistrableBeneficialOwner.post);
-
-router.route(config.UPDATE_REVIEW_STATEMENT_BEFORE_SUBMITTING_URL)
-  .all(
-    authentication,
-    companyAuthentication,
-    navigation.hasOverseasEntity
-  )
-  .get(updateReviewStatement.get);
-// .post(...validator.registrableBeneficialOwner, checkValidations, noChangeRegistrableBeneficialOwner.post);
+  .get(updateReviewStatement.get)
+  .post(updateReviewStatement.post);
 
 router.route(config.UPDATE_CONTINUE_WITH_SAVED_FILING_URL)
   .all(authentication)
